Add a command to add systems to an organization automatically
